% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{blockdiag}
\alias{blockdiag}
\title{Diagonal merging of matrices}
\usage{
blockdiag(...)
}
\arguments{
\item{...}{One or several object of type \code{matrix}}
}
\value{
A matrix obtained by merging the input matrices diagonally
}
\description{
Diagonal merging of matrices
}
\keyword{internal}
