% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{build_D_mat}
\alias{build_D_mat}
\title{Build Difference Matrix of Given Order}
\usage{
build_D_mat(n, q)
}
\arguments{
\item{n}{Number of observations in the problem}

\item{q}{Order of the penalization matrix}
}
\value{
The difference matrix of order \code{q} for a vector containing
\code{n} observations.
}
\description{
Build Difference Matrix of Given Order
}
\keyword{internal}
