% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_deviance}
\alias{compute_deviance}
\title{Deviance for Poisson GLM}
\usage{
compute_deviance(D, D_hat)
}
\arguments{
\item{D}{Vector or matrix containing the number of observed events}

\item{D_hat}{Vector or matrix containing the number of predicted events}
}
\value{
The model deviance
}
\description{
Deviance for Poisson GLM
}
