% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compute_res_deviance}
\alias{compute_res_deviance}
\title{Deviance Residuals for Poisson GLM}
\usage{
compute_res_deviance(D, D_hat)
}
\arguments{
\item{D}{Vector or matrix containing the number of observed events}

\item{D_hat}{Vector or matrix containing the number of predicted events}
}
\value{
A vector or matrix (depending on the input type, will be a matrix
if at least one of the input is) containing the deviance residuals
}
\description{
Deviance Residuals for Poisson GLM
}
\keyword{internal}
