% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compute_XZ_mat}
\alias{compute_XZ_mat}
\title{Compute Block Kronecket Product Matrix}
\usage{
compute_XZ_mat(XZ)
}
\arguments{
\item{XZ}{A list whose components are list of 2 matrices}
}
\value{
A matrix obtained by applying the Kronecker product to the 2 matrices
contained in each element of \code{XZ} and then horizontally concatenating
the result
}
\description{
Compute Block Kronecket Product Matrix
}
\keyword{internal}
