% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{edf_par_to_matrix}
\alias{edf_par_to_matrix}
\title{Diagnosis for Model Fit}
\usage{
edf_par_to_matrix(edf_par, p, q)
}
\arguments{
\item{edf_par}{A vector containing effective degrees of freedom by parameter
with fixed effects in front, fixed-random combinations, then random-random
combinations}

\item{p}{Vector of number of parameters kept on each dimension}

\item{q}{Vector of penalization order on each dimension}
}
\value{
A matrix containing effective degrees of freedom by parameter at the
right position
}
\description{
Diagnosis for Model Fit
}
\keyword{internal}
