% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{eigen_dec}
\alias{eigen_dec}
\title{Eigen decomposition of penalization matrix}
\usage{
eigen_dec(n, q, p)
}
\arguments{
\item{n}{Number of observations in the problem}

\item{q}{Order of the penalization matrix}

\item{p}{Number of eigenvectors to keep}
}
\value{
A list with components : - \code{U} a matrix whose columns are the
eigenvectors of the penalization matrix - \code{s} a vector containing the
eigenvalues sorted in ascending order
}
\description{
Eigen decomposition of penalization matrix
}
\keyword{internal}
