% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_diagnosis}
\alias{get_diagnosis}
\title{Diagnosis for Model Fit}
\usage{
get_diagnosis(dev, pen, sum_edf, n_pos, p, q, s, lambda, R, l_sat = 0)
}
\arguments{
\item{dev}{Deviance of the model}

\item{pen}{Penalization force}

\item{sum_edf}{Effective dimension}

\item{n_pos}{Number of strictly positive weights}

\item{p}{Number of parameters}

\item{q}{Orders of the penalization matrices}

\item{s}{eigenvalues for the penalization matrix}

\item{lambda}{smoothing parameters}

\item{R}{triangular factor}

\item{l_sat}{likelihood for the saturated model}
}
\value{
A data.frame containing various diagnosis about the fit
}
\description{
Diagnosis for Model Fit
}
\keyword{internal}
