% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{map}
\alias{map}
\title{Lapply with Custom Return Type}
\usage{
map(x, f, output_type = "list")
}
\arguments{
\item{x}{A vector or list}

\item{f}{A function}

\item{output_type}{The desired return type. Should be one of "list" (the
default), "integer", "numeric" or "character".}
}
\value{
A list (if output_type = "list") or a vector, sharing names with x,
built by applying f to each element of x. This is a very rough
implementation of the map function from the wonderful package purrr meant
to remove the dependency from this package.
}
\description{
Lapply with Custom Return Type
}
\keyword{internal}
