% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.WH_1d}
\alias{plot.WH_1d}
\title{Plot 1D WH fit}
\usage{
\method{plot}{WH_1d}(x, what = "fit", trans, ...)
}
\arguments{
\item{x}{An object returned by the WH_1d function}

\item{what}{What should be plotted (fit, res, edf)}

\item{trans}{An (optional) transformation to be applied to the data}

\item{...}{Not used}
}
\value{
A plot representing the given element of the fit...
}
\description{
Plot 1D WH fit
}
\examples{
keep <- which(portfolios_mort[[1]]$ec > 0)
d <- portfolios_mort[[1]]$d[keep]
ec <- portfolios_mort[[1]]$ec[keep]

WH_1d(d, ec) |> plot()
WH_1d(d, ec) |> plot("res")
WH_1d(d, ec) |> plot("edf")

}
