% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.WH_2d}
\alias{plot.WH_2d}
\title{Plot 2D WH fit}
\usage{
\method{plot}{WH_2d}(x, what = "y_hat", trans, ...)
}
\arguments{
\item{x}{An object of class \code{"WH_2d"} returned by the \code{\link[=WH]{WH()}} function}

\item{what}{What should be plotted (y_hat, std_y_hat, res, edf)}

\item{trans}{An (optional) transformation to be applied to the data}

\item{...}{Not used}
}
\value{
A plot representing the given element of the fit...
}
\description{
Plot 2D WH fit
}
\examples{
d  <- portfolio_LTC$d
ec <- portfolio_LTC$ec

WH(d, ec) |> plot()
WH(d, ec) |> plot("std_y_hat")

}
