% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{data}
\name{portfolio_LTC}
\alias{portfolio_LTC}
\title{Agregated Long-Term Care Dataset}
\format{
A dataset obtaining by agregating the information from a fictive
long-term care annuitant database with 5,000 annuitants over a 10-year
observation period. The dataset is supplied as a list with two components :
\describe{\item{exit}{A matrix containing the number of observed deaths for
each combination of age and duration in LTC where at least one death
has been observed} \item{expo}{A matrix containing the associated central
exposure in person-years for each combination of age and duration in LTC in d}}
}
\usage{
portfolio_LTC
}
\description{
Agregated dataset built from a synthetic long-term care portfolio
}
\keyword{datasets}
