% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{data}
\name{portfolio_mort}
\alias{portfolio_mort}
\title{Agregated Mortality Dataset}
\format{
An agregated dataset agregating the information from an annuity
portfolio with 100,000 contributors on a 10-year observation period. The
dataset is supplied as a list with two components : \describe{\item{exit}{A
vector containing the portfolio number of observed deaths for each age from
0 to 120} \item{expo}{A vector containing the portfolio central exposure in
person-years for each age from 0 to 119}}
}
\usage{
portfolio_mort
}
\description{
Agregated dataset built from a simulated mortality portfolio
}
\keyword{datasets}
