% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{data}
\name{portfolios_LTC}
\alias{portfolios_LTC}
\title{Simulated Long-Term Care Datasets}
\format{
A list of 3 agregated datasets obtained from fictive LTC annuity
portfolios of increasing sizes, each one containing :
\describe{\item{exit}{A matrix containing the portfolio number of observed
deaths for each combination of age and duration in LTC} \item{expo}{A
matrix containing the portfolio central exposure in person-years for each
combination of age and duration in LTC}}
}
\usage{
portfolios_LTC
}
\description{
Agregated long-term care datasets built from simulated portfolios
}
\keyword{datasets}
