% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{data}
\name{portfolios_mort}
\alias{portfolios_mort}
\title{Simulated Mortality Datasets}
\format{
A list of 3 agregated datasets obtained from fictive annuity
portfolios of increasing sizes, each one containing :
\describe{\item{exit}{A vector containing the portfolio number of observed
deaths for each age} \item{expo}{A vector containing the portfolio central
exposure in person-years for each age}}
}
\usage{
portfolios_mort
}
\description{
Agregated mortality datasets built from simulated portfolios
}
\keyword{datasets}
