% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.WH_1d}
\alias{predict.WH_1d}
\title{Predict new values using a fitted 1D WH model}
\usage{
\method{predict}{WH_1d}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object returned by the WH_1d function}

\item{newdata}{A list containing a vector indicating the new observation
positions}

\item{...}{Not used}
}
\value{
An object of class \code{WH_1d} with additional components for model
prediction.
}
\description{
Extrapolate the model for new values of the covariates
}
\examples{
keep <- which(portfolios_mort[[1]]$ec > 0)
d <- portfolios_mort[[1]]$d[keep]
ec <- portfolios_mort[[1]]$ec[keep]

WH_1d(d, ec) |> predict(newdata = 18:99) |> plot()

}
