% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.WH_2d}
\alias{predict.WH_2d}
\title{Predict 2D WH values}
\usage{
\method{predict}{WH_2d}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object returned by one of the WH 1D fitting function}

\item{newdata}{A list containing two vectors indicating the new observation
positions}

\item{...}{Not used}
}
\value{
An WH_2d object with predicted values...
}
\description{
Predict 2D WH values
}
