% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict_WH_1d_old}
\alias{predict_WH_1d_old}
\title{Predict new values using a fitted 1D WH model}
\usage{
predict_WH_1d_old(object, newdata = NULL, unconstrained = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{"WH_1d"} returned by the \code{\link[=WH_1d]{WH_1d()}} function}

\item{newdata}{A vector containing the new observation positions}

\item{unconstrained}{Should the unconstrained (approximate) solution be also
computed ? Only used to justify the need for a constrained solution.}

\item{...}{Not used}
}
\value{
An object of class \code{"WH_1d"} with additional components for model
prediction.
}
\description{
Extrapolate the model for new observation. There is now a simpler version of
this function.
}
\examples{
keep <- which(portfolio_mort$ec > 0)
d <- portfolio_mort$d[keep]
ec <- portfolio_mort$ec[keep]

WH_1d(d, ec) |> predict(newdata = 18:99) |> plot()

}
