% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.WH_1d}
\alias{print.WH_1d}
\title{Display of 1D WH object}
\usage{
\method{print}{WH_1d}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"WH_1d"} returned by the \code{\link[=WH]{WH()}} function}

\item{...}{Not used}
}
\value{
Invisibly returns \code{x}.
}
\description{
Display of 1D WH object
}
\examples{
d <- portfolio_mort$d
ec <- portfolio_mort$ec

WH(d, ec) |> summary()

}
