% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary.WH_1d}
\alias{summary.WH_1d}
\title{Summary of 1D WH object}
\usage{
\method{summary}{WH_1d}(object, ...)
}
\arguments{
\item{object}{An object returned by the WH_1d function}

\item{...}{Not used}
}
\value{
A plot representing the given element of the fit...
}
\description{
Summary of 1D WH object
}
\examples{
keep <- which(portfolios_mort[[1]]$ec > 0)
d <- portfolios_mort[[1]]$d[keep]
ec <- portfolios_mort[[1]]$ec[keep]

y <- log(d / ec)
y[d == 0] <- - 20
wt <- d

WH_1d(d, ec) |> summary()
WH_1d(d, ec, method = "optim") |> summary()
WH_1d(d, ec, criterion = "GCV") |> summary()
WH_1d(y = y, wt = wt) |> summary()

}
