% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary.WH_2d}
\alias{summary.WH_2d}
\title{Summary of 2D WH object}
\usage{
\method{summary}{WH_2d}(object, ...)
}
\arguments{
\item{object}{An object returned by the WH_2d function}

\item{...}{Not used}
}
\value{
A plot representing the given element of the fit...
}
\description{
Summary of 2D WH object
}
\examples{
keep_age <- which(rowSums(portfolios_LTC[[1]]$ec) > 1e2)
keep_duration <- which(colSums(portfolios_LTC[[1]]$ec) > 1e2)

d  <- portfolios_LTC[[1]]$d[keep_age, keep_duration]
ec <- portfolios_LTC[[1]]$ec[keep_age, keep_duration]

WH_2d(d, ec) |> summary()

}
