% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{package}
\name{WH-package}
\alias{WH-package}
\title{WH : Enhanced Implementation of Whittaker-Henderson Smoothing}
\description{
An enhanced implementation of Whittaker-Henderson smoothing for the gradation
of one-dimensional and two-dimensional actuarial tables used to quantify Life
Insurance risks. \code{WH} is based on the methods described in Biessy (2023)
\url{doi:10.48550/arXiv.2306.06932}. Among other features, it generalizes the
original smoothing algorithm to maximum likelihood estimation, automatically
selects the smoothing parameter(s) and extrapolates beyond the range of data.
}
