% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_1d_ml_fs}
\alias{WH_1d_ml_fs}
\title{Whittaker-Henderson Smoothing (Maximum Likelihood, Generalized Fellner-Schall update)}
\usage{
WH_1d_ml_fs(
  d,
  ec,
  p = length(d),
  q = 2,
  lambda = 1000,
  verbose = FALSE,
  accu_edf = 1e-10,
  accu_dev = 1e-12
)
}
\arguments{
\item{d}{Vector of observed events}

\item{ec}{Vector of central exposure}

\item{p}{The number of eigenvectors to keep}

\item{q}{Order of penalization. Polynoms of degrees q - 1 are considered
smooth and are therefore unpenalized}

\item{lambda}{Initial smoothing parameter}

\item{verbose}{Should information about the optimization progress be
displayed}

\item{accu_edf}{Tolerance for the convergence of the outer optimization
procedure}

\item{accu_dev}{Tolerance for the convergence of the optimization procedure}
}
\value{
A list containing model fit, residuals and associated uncertainty.
Smoothing parameter is chosen using the generalized Fellner-Schall update
formula
}
\description{
Whittaker-Henderson Smoothing (Maximum Likelihood, Generalized Fellner-Schall update)
}
