% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_1d_reg_fixed_lambda}
\alias{WH_1d_reg_fixed_lambda}
\title{Whittaker-Henderson Smoothing (Regression, fixed lambda)}
\usage{
WH_1d_reg_fixed_lambda(
  y,
  wt = rep(1, length(y)),
  lambda = 1000,
  p = length(y),
  q = 2
)
}
\arguments{
\item{y}{Vector of observations}

\item{wt}{Optional vector of weights}

\item{lambda}{Smoothing parameter}

\item{p}{The number of eigenvectors to keep}

\item{q}{Order of penalization. Polynoms of degrees q - 1 are considered
smooth and are therefore unpenalized}
}
\value{
A list containing model fit, residuals and associated uncertainty
}
\description{
Whittaker-Henderson Smoothing (Regression, fixed lambda)
}
