% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_1d_reg_fs}
\alias{WH_1d_reg_fs}
\title{Whittaker-Henderson Smoothing (Regression, Generalized Fellner-Schall update)}
\usage{
WH_1d_reg_fs(
  y,
  wt = rep(1, length(y)),
  p = length(y),
  q = 2,
  lambda = 1000,
  verbose = FALSE,
  accu_edf = 1e-10
)
}
\arguments{
\item{y}{Vector of observations}

\item{wt}{Optional vector of weights}

\item{p}{The number of eigenvectors to keep}

\item{q}{Order of penalization. Polynoms of degrees q - 1 are considered
smooth and are therefore unpenalized}

\item{lambda}{Initial smoothing parameter}

\item{verbose}{Should information about the optimization progress be
displayed}

\item{accu_edf}{Tolerance for the convergence of the outer optimization
procedure}
}
\value{
An object of class \code{"WH_1d"} i.e. a list containing model fit,
variance, residuals and degrees of freedom as well as diagnosis to asses
the quality of the fit.
}
\description{
Whittaker-Henderson Smoothing (Regression, Generalized Fellner-Schall update)
}
\keyword{internal}
