% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_1d_reg_optim}
\alias{WH_1d_reg_optim}
\title{Whittaker-Henderson Smoothing (Regression, optimize function)}
\usage{
WH_1d_reg_optim(
  y,
  wt = rep(1, length(y)),
  p = length(y),
  q = 2,
  criterion = "REML",
  lambda = 1000,
  verbose = FALSE,
  accu_edf = 1e-10
)
}
\arguments{
\item{y}{Vector of observations}

\item{wt}{Optional vector of weights}

\item{p}{The number of eigenvectors to keep}

\item{q}{Order of penalization. Polynoms of degrees q - 1 are considered
smooth and are therefore unpenalized}

\item{criterion}{Criterion used to choose the smoothing parameter. One of
"GCV" (default), "AIC" or "BIC".}

\item{lambda}{Initial smoothing parameter}

\item{verbose}{Should information about the optimization progress be
displayed}

\item{accu_edf}{Tolerance for the convergence of the outer optimization
procedure}
}
\value{
A list containing model fit, residuals and associated uncertainty.
  Smoothing parameter is chosen through successive fits, using the
  \code{optimize} function
}
\description{
Whittaker-Henderson Smoothing (Regression, optimize function)
}
