% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_2d}
\alias{WH_2d}
\title{1D Whittaker-Henderson Smoothing}
\usage{
WH_2d(
  d,
  ec,
  lambda,
  method,
  p,
  max_dim = 1000,
  q = c(2, 2),
  framework,
  y,
  wt,
  ...
)
}
\arguments{
\item{d}{Matrix of observed events whose rows and columns should be named.
Required in case of maximum likelihood estimation}

\item{ec}{Matrix of central exposure. Required in case of maximum likelihood
estimation. The central exposure corresponds to the sum of the exposure
period over the insured population. An individual experiencing an event of
interest during the year will no longer be exposed afterward and the
exposure should be reduced accordingly.}

\item{lambda}{Smoothing parameter vector of size \code{2}. Optional. If
missing, an optimization procedure will be used to find the optimal
smoothing parameter. If provided, no optimal smoothing parameter search
will be performed unless the \code{method} argument is also provided, in
which case \code{lambda} will be used as the starting parameter for the
optimization procedure.}

\item{method}{The method to be used to find an optimal smoothing parameter.
Default to `"fixed_lambda"` if \code{lambda} is supplied, meaning no
optimization is performed. Otherwise, default to `"fs"` which means the
generalized Fellner-Schall method is used. The third option is `"optim"`
meaning the \code{optim} function from package \code{stats} will be used
with the default Nelder-Mead algorithm. The optimization criterion used for
the `"fs"` method as well as the default of the `"optim"` method is is the
restricted maximum likelihood (REML). For the `"optim"` method others
options such as `AIC`, `BIC` and `GCV` may be used by supplying for example
\code{criterion = "AIC"} as an extra argument to the function}

\item{p}{Optional vector of size \code{2}. Maximum number of eigenvectors to
keep on each dimension after performing the eigen decomposition of the
penalization matrix. If missing, will be automatically computed so that the
dimensions of (square) matrices involved in the optimization problem
remains lower to the \code{max_dim} argument}

\item{max_dim}{Maximal number of rows (or equivalently columns) of the square
matrices that should be involved in the optimization problem. Default is
\code{1000}. Values higher than \code{2000} may result in issues with both
memory usage and computation times.}

\item{q}{Order of penalization vector of size \code{2}. Polynoms of degrees
\code{(q[[1]] - 1,q[[2]] - 1)} are considered smooth and are therefore
unpenalized. Should be left to the default of \code{c(2,2)} for most
practical applications.}

\item{framework}{Default framework is `"ml"` which stands for maximum
likelihood unless both the \code{y} and \code{wt} arguments are provided,
in which case the `"reg"` or regression framework is used. The regression
framework is an asymptotical approximation of the maximum likelihood
likelihood framework.}

\item{y}{Matrix of observations whose rows and columns should be named. Used
only in the regression framework and even in this case will be
automatically computed if the \code{d} and \code{ec} arguments are
provided. May be useful when using Whittaker-Henderson smoothing outside of
the survival analysis framework.}

\item{wt}{Matrix of weights. As for the observation vector \code{y}, used
only in the regression framework and even in this case will be
automatically computed if the \code{d} and \code{ec} arguments are
provided. May be useful when using Whittaker-Henderson smoothing outside of
the survival analysis framework.}

\item{...}{Additional parameters passed to the smoothing function called.}
}
\value{
A fitted \code{WH_2d} object.
}
\description{
1D Whittaker-Henderson Smoothing
}
