% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_2d_fixed_lambda}
\alias{WH_2d_fixed_lambda}
\title{2D Whittaker-Henderson Smoothing (Maximum Likelihood, fixed lambda)}
\usage{
WH_2d_fixed_lambda(
  d,
  ec,
  y,
  wt,
  lambda = c(1000, 1000),
  q = c(2, 2),
  p,
  reg = FALSE,
  verbose = FALSE,
  accu_dev = 1e-12
)
}
\arguments{
\item{y}{Matrix of observations}

\item{wt}{Optional matrix of weights}

\item{lambda}{Vector of smoothing parameter}

\item{q}{Matrix of orders of penalization. Polynoms of degrees q - 1 are considered
smooth and are therefore unpenalized}

\item{p}{The number of eigenvectors to keep on each dimension}
}
\value{
An object of class \code{"WH_2d"} i.e. a list containing model fit,
variance, residuals and degrees of freedom as well as diagnosis to asses
the quality of the fit.
}
\description{
2D Whittaker-Henderson Smoothing (Maximum Likelihood, fixed lambda)
}
\keyword{internal}
