% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_2d_ml_fixed_lambda}
\alias{WH_2d_ml_fixed_lambda}
\title{2D Whittaker-Henderson Smoothing (Maximum Likelihood, fixed lambda)}
\usage{
WH_2d_ml_fixed_lambda(
  d,
  ec,
  lambda = c(1000, 1000),
  p = dim(d),
  q = c(2, 2),
  verbose = FALSE,
  accu_dev = 1e-12
)
}
\arguments{
\item{d}{Vector of observed events}

\item{ec}{Vector of central exposure}

\item{lambda}{Vector of smoothing parameter}

\item{p}{The number of eigenvectors to keep on each dimension}

\item{q}{Matrix of orders of penalization. Polynoms of degrees q - 1 are considered
smooth and are therefore unpenalized}

\item{verbose}{Should information about the optimization progress be
displayed}

\item{accu_dev}{Tolerance for the convergence of the optimization procedure}
}
\value{
A list containing model fit, residuals and associated uncertainty
}
\description{
2D Whittaker-Henderson Smoothing (Maximum Likelihood, fixed lambda)
}
