% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_2d_ml_fs}
\alias{WH_2d_ml_fs}
\title{2D Whittaker-Henderson Smoothing (Maximum Likelihood, Generalized Fellner-Schall update)}
\usage{
WH_2d_ml_fs(
  d,
  ec,
  q = c(2, 2),
  p = dim(d),
  verbose = FALSE,
  lambda = c(1000, 1000),
  accu_edf = 1e-10,
  accu_dev = 1e-12
)
}
\arguments{
\item{d}{Vector of observed events}

\item{ec}{Vector of central exposure}

\item{q}{Matrix of orders of penalization. Polynoms of degrees q - 1 are considered
smooth and are therefore unpenalized}

\item{p}{The number of eigenvectors to keep on each dimension}

\item{verbose}{Should information about the optimization progress be
displayed}

\item{lambda}{Vector of smoothing parameter}

\item{accu_edf}{Tolerance for the convergence of the outer optimization
procedure}

\item{accu_dev}{Tolerance for the convergence of the optimization procedure}
}
\value{
A list containing model fit, residuals and associated uncertainty.
Smoothing parameter is chosen using the generalized Fellner-Schall update
formula
}
\description{
2D Whittaker-Henderson Smoothing (Maximum Likelihood, Generalized Fellner-Schall update)
}
