% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_2d_reg_fixed_lambda}
\alias{WH_2d_reg_fixed_lambda}
\title{2D Whittaker-Henderson Smoothing (Regression, fixed lambda)}
\usage{
WH_2d_reg_fixed_lambda(
  y,
  wt = matrix(1, nrow = nrow(y), ncol = ncol(y)),
  lambda = c(1000, 1000),
  p = dim(y),
  q = c(2, 2)
)
}
\arguments{
\item{y}{Matrix of observations}

\item{wt}{Optional matrix of weights}

\item{lambda}{Vector of smoothing parameter}

\item{p}{The number of eigenvectors to keep on each dimension}

\item{q}{Matrix of orders of penalization. Polynoms of degrees q - 1 are considered
smooth and are therefore unpenalized}
}
\value{
An object of class \code{"WH_2d"} i.e. a list containing model fit,
variance, residuals and degrees of freedom as well as diagnosis to asses
the quality of the fit.
}
\description{
2D Whittaker-Henderson Smoothing (Regression, fixed lambda)
}
