% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_2d_reg_fs}
\alias{WH_2d_reg_fs}
\title{2D Whittaker-Henderson Smoothing (Regression, Fellner-Schall update)}
\usage{
WH_2d_reg_fs(
  y,
  wt = matrix(1, nrow = nrow(y), ncol = ncol(y)),
  q = c(2, 2),
  p = dim(y),
  lambda = c(1000, 1000),
  verbose = FALSE,
  accu_edf = 1e-10
)
}
\arguments{
\item{y}{Vector of observations}

\item{wt}{Optional vector of weights}

\item{q}{Order of penalization. Polynoms of degrees q - 1 are considered
smooth and are therefore unpenalized}

\item{p}{The number of eigenvectors to keep}

\item{lambda}{Initial smoothing parameters}

\item{verbose}{Should information about the optimization progress be
displayed}

\item{accu_edf}{Tolerance for the convergence of the outer optimization
procedure}
}
\value{
An object of class \code{"WH_2d"} i.e. a list containing model fit,
variance, residuals and degrees of freedom as well as diagnosis to asses
the quality of the fit.
}
\description{
2D Whittaker-Henderson Smoothing (Regression, Fellner-Schall update)
}
\keyword{internal}
