% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{blockdiag}
\alias{blockdiag}
\title{Diagonal Stacking of Matrices}
\usage{
blockdiag(...)
}
\arguments{
\item{...}{One or several object of type \code{matrix}}
}
\value{
A matrix obtained by stacking the input matrices diagonally
}
\description{
Diagonal Stacking of Matrices
}
\keyword{internal}
