% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{build_D_mat}
\alias{build_D_mat}
\title{Compute Block Kronecket Product Matrix}
\usage{
build_D_mat(n, q)
}
\arguments{
\item{n}{Number of observations in the problem}

\item{q}{Order of the penalization matrix}
}
\value{
Difference matrix of order \code{q} for a vector containing \code{n}
observations.
}
\description{
Compute Block Kronecket Product Matrix
}
\keyword{internal}
