% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eigen_dec}
\alias{eigen_dec}
\title{Eigen Decomposition of Penalization Matrix}
\usage{
eigen_dec(n, q, p)
}
\arguments{
\item{n}{Number of observations in the problem}

\item{q}{Order of the penalization matrix}

\item{p}{Number of eigenvectors to keep}
}
\value{
A list with components : - \code{X} a matrix whose columns are the
eigenvectors associated with 0 eigenvalues - \code{Z} a matrix whose
columns are the eigenvectors associated with non-0 eigenvalues sorted in
ascending order - \code{s} a vector containing the eigenvalues sorted in
ascending order
}
\description{
Eigen Decomposition of Penalization Matrix
}
