% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_diagnosis}
\alias{get_diagnosis}
\title{Diagnosis for Model Fit}
\usage{
get_diagnosis(dev, pen, sum_edf, n_pos, tr_log_P, tr_log_Psi)
}
\arguments{
\item{dev}{Deviance of the model}

\item{pen}{Penalization force}

\item{sum_edf}{Effective dimension}

\item{n_pos}{Number of strictly positive weights}

\item{tr_log_P}{Trace of logarithm of penalization matrix}

\item{tr_log_Psi}{Trace of logarithm of variance-covariance matrix}
}
\value{
A data.frame containing various diagnosis about the fit
}
\description{
Diagnosis for Model Fit
}
\keyword{internal}
