% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.WH_1d}
\alias{plot.WH_1d}
\title{Plot 1D WH fit}
\usage{
\method{plot}{WH_1d}(x, what = "fit", trans)
}
\arguments{
\item{x}{An object returned by one of the WH 1D fitting function}

\item{what}{What should be plotted (fit, res, edf)}

\item{trans}{An (optional) transformation to be applied to the data}
}
\value{
A plot representing the given element of the fit...
}
\description{
Plot 1D WH fit
}
