% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.WH_2d}
\alias{plot.WH_2d}
\title{Plot 1D WH fit}
\usage{
\method{plot}{WH_2d}(x, what = "y_hat", trans)
}
\arguments{
\item{x}{An object returned by one of the WH 2D fitting function}

\item{what}{What should be plotted (y_hat, std_y_hat, res, edf)}

\item{trans}{An (optional) transformation to be applied to the data}
}
\value{
A plot representing the given element of the fit...
}
\description{
Plot 1D WH fit
}
