% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.WH_2d}
\alias{plot.WH_2d}
\title{Plot 2D WH fit}
\usage{
\method{plot}{WH_2d}(x, what = "y_hat", trans, ...)
}
\arguments{
\item{x}{An object of class \code{"WH_2d"} returned by the \code{\link[=WH_2d]{WH_2d()}} function}

\item{what}{What should be plotted (y_hat, std_y_hat, res, edf)}

\item{trans}{An (optional) transformation to be applied to the data}

\item{...}{Not used}
}
\value{
A plot representing the given element of the fit...
}
\description{
Plot 2D WH fit
}
\examples{

keep_age <- which(rowSums(portfolios_LTC[[1]]$ec) > 1e2)
keep_duration <- which(colSums(portfolios_LTC[[1]]$ec) > 1e2)

d  <- portfolios_LTC[[1]]$d[keep_age, keep_duration]
ec <- portfolios_LTC[[1]]$ec[keep_age, keep_duration]

WH_2d(d, ec) |> plot()
WH_2d(d, ec) |> plot("std_y_hat")

}
