% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{data}
\name{portfolio_LTC}
\alias{portfolio_LTC}
\title{Agregated Long-Term Care Dataset}
\format{
An agregated dataset obtaining by agregating the information from a
fictive annuitant database from a long-term care cover with 5,000
annuitants on a 10-year observation period. The dataset is supplied as a
list with two components : \describe{\item{exit}{A matrix
containing the portfolio number of observed deaths for each combination of
age from 0 to 119 and duration in LTC from 0 to 29} \item{expo}{A matrix
containing the portfolio central exposure in person-years for each
combination of age from 0 to 119 and duration in LTC from 0 to 29}}
}
\usage{
portfolio_LTC
}
\description{
Agregated dataset built from a simulated long-term care portfolio
}
\keyword{datasets}
