% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{data}
\name{portfolio_LTC}
\alias{portfolio_LTC}
\title{Agregated Long-Term Care Dataset}
\format{
An synthetic aggregated dataset with death and exposure counts from a
simulated long-term care portfolio with 100,000 contributors on a 20-year
observation period (only deaths following long-term care claims are
counted). The dataset is supplied as a
list with two components : \describe{\item{d}{A matrix
containing the portfolio number of observed deaths for each combination of
age from 70 to 100 (excluded) and duration in LTC from 0 to 15 (excluded)} \item{ec}{A matrix
containing the portfolio central exposure in person-years for each
combination age from 70 to 100 (excluded) and duration in LTC from 0 to 15 (excluded)}}
}
\usage{
portfolio_LTC
}
\description{
Agregated dataset built from a simulated long-term care portfolio
}
\keyword{datasets}
