% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{data}
\name{portfolio_mort}
\alias{portfolio_mort}
\title{Agregated Mortality Dataset}
\format{
A dataset containing the information from a simulated annuity
portfolio with 100,000 contributors over a 10-year observation period. The
dataset is supplied as a list with two components : \describe{\item{d}{A
vector containing the number of observed deaths for ages where at
least one death has been observed} \item{ec}{A vector containing the associated
central exposure in person-years for each age in d}}
}
\usage{
portfolio_mort
}
\description{
Agregated dataset built from a synthetic mortality portfolio
}
\keyword{datasets}
