% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{data}
\name{portfolio_mort}
\alias{portfolio_mort}
\title{Aggregated Mortality Dataset}
\format{
An synthetic aggregated dataset with death and exposure counts from a
simulated annuity portfolio with 100,000 contributors on a 20-year
observation period. The
dataset is supplied as a list with two components : \describe{\item{d}{A
vector containing the portfolio number of observed deaths for each age from
50 to 95 (excluded)} \item{ec}{A vector containing the portfolio central exposure in
person-years for each age from 50 to 95 (excluded)}}
}
\usage{
portfolio_mort
}
\description{
Aggregated dataset built from a simulated mortality portfolio
}
\keyword{datasets}
