% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{data}
\name{portfolios_LTC}
\alias{portfolios_LTC}
\title{Simulated Long-Term Care Datasets}
\format{
A list of 3 datasets of increasing sizes, each one containing :
\describe{ \item{expo}{An array containing the portfolio central exposure
in years for each combination of age, duration in LTC and gender}
\item{exit}{An array containing the number of observed deaths for each
combination of age, duration in LTC and gender}}
}
\usage{
portfolios_LTC
}
\description{
Agregated long-term care datasets built from simulated portfolios
}
\keyword{datasets}
