% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{data}
\name{portfolios_mort}
\alias{portfolios_mort}
\title{Simulated Mortality Datasets}
\format{
A list of 3 datasets of increasing sizes, each one containing :
\describe{ \item{expo}{An array containing the portfolio central exposure
in years for each combination of age and gender} \item{exit}{An array
containing the number of observed deaths for each combination of age and
gender}}
}
\usage{
portfolios_mort
}
\description{
Agregated mortality datasets built from simulated portfolios
}
\keyword{datasets}
