% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.WH_1d}
\alias{predict.WH_1d}
\title{Predict 1D WH values}
\usage{
\method{predict}{WH_1d}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object returned by one of the WH 1D fitting function}

\item{newdata}{A list containing a vector indicating the new observation
positions}

\item{...}{Not used}
}
\value{
An WH_1d object with predicted values...
}
\description{
Predict 1D WH values
}
