% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.WH_1d}
\alias{predict.WH_1d}
\title{Predict new values using a fitted 1D WH model}
\usage{
\method{predict}{WH_1d}(object, newdata = NULL, unconstrained = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{"WH_1d"} returned by the \code{\link[=WH_1d]{WH_1d()}} function}

\item{newdata}{A list containing a vector indicating the new observation
positions}

\item{unconstrained}{Should the unconstrained (approximate) solution be also
computed ? Only used to justify the use for a constrained solution.}

\item{...}{Not used}
}
\value{
An object of class \code{"WH_1d"} with additional components for model
prediction.
}
\description{
Extrapolate the model for new values of the covariates
}
\examples{
keep <- which(portfolio_mort$ec > 0)
d <- portfolio_mort$d[keep]
ec <- portfolio_mort$ec[keep]

WH_1d(d, ec) |> predict(newdata = 18:99) |> plot()

}
