% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.WH_1d}
\alias{predict.WH_1d}
\title{Predict new values using a fitted 1D WH model}
\usage{
\method{predict}{WH_1d}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"WH_1d"} returned by the \code{\link[=WH]{WH()}} function}

\item{newdata}{A vector containing the position of new observations.
Observations from the fit will automatically be added to this, in the
adequate order}

\item{...}{Not used}
}
\value{
An object of class \code{"WH_1d"} with additional components for model
prediction.
}
\description{
Extrapolate the model for new observations.
}
\examples{

object <- WH(portfolio_mort$d, portfolio_mort$ec)
object_extra <- predict(object, newdata = 18:99)
plot(object_extra)

}
