% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.WH_2d}
\alias{predict.WH_2d}
\title{Predict new values using a fitted 2D WH model}
\usage{
\method{predict}{WH_2d}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"WH_2d"} returned by the \code{\link[=WH]{WH()}} function}

\item{newdata}{A list containing two vectors indicating the new observation
positions}

\item{...}{Not used}
}
\value{
An object of class \code{"WH_2d"} with additional components for model
prediction.
}
\description{
Extrapolate the model for new observations in a way that is consistent with
the fitted values
}
\examples{
object <- WH(portfolio_LTC$d, portfolio_LTC$ec)
object_extra <- predict(object, newdata = list(age = 60:109, duration = 0:19))
plot(object_extra)

}
