% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.WH_2d}
\alias{predict.WH_2d}
\title{Predict new values using a fitted 2D WH model}
\usage{
\method{predict}{WH_2d}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object returned by the WH_2d function}

\item{newdata}{A list containing two vectors indicating the new observation
positions}

\item{...}{Not used}
}
\value{
An object of class \code{WH_2d} with additional components for model
prediction.
}
\description{
Extrapolate the model for new values of the covariates in a way that is
consistent with the fitted values
}
\examples{
keep_age <- which(rowSums(portfolios_LTC[[1]]$ec) > 1e2)
keep_duration <- which(colSums(portfolios_LTC[[1]]$ec) > 1e2)

d  <- portfolios_LTC[[1]]$d[keep_age, keep_duration]
ec <- portfolios_LTC[[1]]$ec[keep_age, keep_duration]

WH_2d(d, ec) |> predict(newdata = list(age = 50:99, duration = 0:19)) |> plot()

}
