% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.WH_2d}
\alias{predict.WH_2d}
\title{Predict new values using a fitted 2D WH model}
\usage{
\method{predict}{WH_2d}(object, newdata = NULL, n_coef = 10, ...)
}
\arguments{
\item{object}{An object of class \code{"WH_2d"} returned by the \code{\link[=WH_2d]{WH_2d()}} function}

\item{newdata}{A list containing two vectors indicating the new observation
positions}

\item{n_coef}{Number of additional coefficients to be added on each side of
the fit to ensure a smooth transition. If this is lower than the number of
new observations, will provide an approximate extrapolation but much
faster. Default of \code{10} should already be pretty accurate.}

\item{...}{Not used}
}
\value{
An object of class \code{"WH_2d"} with additional components for model
prediction.
}
\description{
Extrapolate the model for new observations in a way that is consistent with
the fitted values
}
\examples{
keep_age <- which(rowSums(portfolio_LTC$ec) > 1e2)
keep_duration <- which(colSums(portfolio_LTC$ec) > 1e2)

d  <- portfolio_LTC$d[keep_age, keep_duration]
ec <- portfolio_LTC$ec[keep_age, keep_duration]

WH_2d(d, ec) |> predict(newdata = list(age = 50:99, duration = 0:19)) |> plot()

}
