% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.WH_1d}
\alias{print.WH_1d}
\title{Display of 1D WH object}
\usage{
\method{print}{WH_1d}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"WH_1d"} returned by the \code{\link[=WH_1d]{WH_1d()}} function}

\item{...}{Not used}
}
\value{
Invisibly returns \code{x}.
}
\description{
Display of 1D WH object
}
\examples{
keep <- which(portfolios_mort[[1]]$ec > 0)
d <- portfolios_mort[[1]]$d[keep]
ec <- portfolios_mort[[1]]$ec[keep]

y <- log(d / ec)
y[d == 0] <- - 20
wt <- d

WH_1d(d, ec) |> summary()
WH_1d(d, ec, method = "optim") |> summary()
WH_1d(d, ec, criterion = "GCV") |> summary()
WH_1d(y = y, wt = wt) |> summary()

}
