% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.WH_2d}
\alias{print.WH_2d}
\title{Display of 2D WH object}
\usage{
\method{print}{WH_2d}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"WH_2d"} returned by the \code{\link[=WH_2d]{WH_2d()}} function}

\item{...}{Not used}
}
\value{
Invisibly returns \code{x}.
}
\description{
Display of 2D WH object
}
\examples{
keep_age <- which(rowSums(portfolio_LTC$ec) > 1e2)
keep_duration <- which(colSums(portfolio_LTC$ec) > 1e2)

d  <- portfolio_LTC$d[keep_age, keep_duration]
ec <- portfolio_LTC$ec[keep_age, keep_duration]

WH_2d(d, ec)

}
