% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{package}
\name{WH-package}
\alias{WH-package}
\title{WH : A Modern Take on Whittaker-Henderson Smoothing}
\description{
This package implements the popular Whittaker-Henderson Smoothing relying on the latest developments.
}
