% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH-package.R
\docType{package}
\name{WH-package}
\alias{WH-package}
\title{WH : a modern take on Whittaker-Henderson smoothing}
\description{
The package implements the popular (at least among actuaries)
Whittaker-Henderson smoothing relying on the latest developments.
}
