% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_1d}
\alias{WH_1d}
\title{1D Whittaker-Henderson Smoothing}
\usage{
WH_1d(
  d,
  ec,
  lambda,
  criterion,
  method,
  q = 2,
  framework,
  y,
  wt,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{d}{Vector of observed events whose elements should be named.}

\item{ec}{Vector of central exposure. The central exposure corresponds to the
sum of the exposure period over the insured population. An individual
experiencing an event of interest during the year will no longer be exposed
afterward and the exposure should be reduced accordingly.}

\item{lambda}{Smoothing parameter. If missing, an optimization procedure will
be used to find the optimal smoothing parameter. If supplied, no optimal
smoothing parameter search will take place unless the \code{criterion} argument
is also supplied, in which case \code{lambda} will be used as the starting
parameter for the optimization procedure.}

\item{criterion}{Criterion to be used for the selection of the optimal
smoothing parameter. Default is \code{"REML"} which stands for restricted
maximum likelihood. Other options include \code{"AIC"}, \code{"BIC"} and \code{"GCV"}.}

\item{method}{Method to be used to find the optimal smoothing parameter.
Default to \code{"fixed_lambda"} if \code{lambda} is supplied, meaning no
optimization is performed. Otherwise, if \code{criterion = "REML"} or the
\code{criterion} argument is missing, default to \code{"fs"} which means the
generalized Fellner-Schall method is used. For other criteria default to
\code{"perf"} meaning the \code{optimize} function from package \code{stats} will be
used.}

\item{q}{Order of penalization. Polynoms of degrees \code{q - 1} are considered
smooth and are therefore unpenalized. Should be left to the default of \code{2}
for most practical applications.}

\item{framework}{Default framework is \code{"ml"} which stands for maximum
likelihood unless the \code{y} argument is also provided, in which case the
\code{"reg"} or regression framework is used. The regression framework is an
asymptotical approximation of the maximum likelihood likelihood framework.}

\item{y}{Optional vector of observations whose elements should be named. Used
only in the regression framework and even in this case will be
automatically computed from the \code{d} and \code{ec} arguments if those are
supplied. May be useful when using Whittaker-Henderson smoothing outside of
the survival analysis framework.}

\item{wt}{Optional vector of weights. As for the observation vector \code{y}, used
only in the regression framework and even in this case will be
automatically computed if the \code{d} argument is supplied. May be useful when
using Whittaker-Henderson smoothing outside of the survival analysis
framework.}

\item{quiet}{Should messages and warnings be displayed ? Default to \code{FALSE},
may be set to \code{TRUE} for experts who want to make numerous calls to this
function.}

\item{...}{Additional parameters passed to the smoothing function called.}
}
\value{
An object of class \code{WH_1d} i.e. a list containing :
\itemize{
\item \code{d} The inputed vector of observed events (if supplied as input)
\item \code{ec} The inputed vector of central exposure (if supplied as input)
\item \code{y} The observation vector (either supplied or computed as y = log(d) - log(ec))
\item \code{wt} The inputed vector of weights (either supplied or computed as \code{d})
\item \code{y_hat} The vector of fitted value
\item \code{std_y_hat} The vector of standard deviation associated with the fitted value
\item \code{res} The vector of model deviance residuals
\item \code{edf} The vector of effective degrees of freedom associated with each observation
\item \code{edf_par} The vector of effective degrees of freedom associated with each eigenvector
\item \code{diagnosis} A data.frame with one line containing the effective degrees of freedom of the model, the deviance of the fit as well as the AIC, BIC, GCV and REML criteria
\item \code{Psi} The variance-covariance matrix associated with the fit, required for
extrapolation.
\item \code{lambda} The smoothing parameter used, either supplied or computed.
\item \code{p} The number of components kept after rank reduction is used (it should not be in the one-dimension case).
\item \code{q} The supplied order for the penalization.
}
}
\description{
Main package function to apply Whittaker-Henderson smoothing in a
unidimensional survival analysis framework. It takes as input a vector of
observed events and a vector of associated central exposure, both depending
on a single covariate, and build a smooth version of the log-hazard rate.
Smoothing parameters may be supplied or automatically chosen according to a
specific criterion such as \code{"REML"} (the default), \code{"AIC"}, \code{"BIC"} or
\code{"GCV"}. Whittaker-Henderson may be applied in a full maximum likelihood
framework or an asymptotic (approximate) gaussian framework.
}
\examples{
keep <- which(portfolio_mort$ec > 0)
d <- portfolio_mort$d[keep]
ec <- portfolio_mort$ec[keep]

y <- log(d / ec)
y[d == 0] <- - 20
wt <- d

# Maximum likelihood
WH_1d(d, ec, lambda = 1e2)
WH_1d(d, ec) # default generalized Fellner-Schall method
WH_1d(d, ec, method = "perf") # alternative method base on optimize function

testthat::expect_equal(WH_1d(d, ec, method = "fs"),
                       WH_1d(d, ec, method = "perf"),
                       tolerance = 1e-1)
# generalized Fellner-Schall method is approximate in maximum likelihood framework

WH_1d(d, ec, criterion = "GCV")
# alternative optimization criteria for smoothing parameter selection

# Regression
WH_1d(y = y, wt = wt, lambda = 1e2)
# regression framework is default when y is supplied
WH_1d(d, y = y, lambda = 1e2) # d is used as wt
WH_1d(d, ec, framework = "reg", lambda = 1e2)
# Setting framework = "reg" forces computation of y from d and ec

WH_1d(y = y, wt = wt)
WH_1d(y = y, wt = wt, method = "perf")
testthat::expect_equal(WH_1d(y = y, wt = wt, method = "fs"),
                       WH_1d(y = y, wt = wt, method = "perf"),
                       tolerance = 1e-6)
# generalized Fellner-Schall method is exact in regression framework

WH_1d(y = y, wt = wt, criterion = "GCV")
# alternative optimization criteria for smoothing parameter selection

}
