% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WH_1d_ml_fixed_lambda}
\alias{WH_1d_ml_fixed_lambda}
\title{Whittaker-Henderson Smoothing (Maximum Likelihood, fixed lambda)}
\usage{
WH_1d_ml_fixed_lambda(
  d,
  ec,
  lambda = 1000,
  p = length(d),
  q = 2,
  verbose = FALSE,
  accu_dev = 1e-12
)
}
\arguments{
\item{d}{Vector of observed events}

\item{ec}{Vector of central exposure}

\item{lambda}{Smoothing parameter}

\item{p}{The number of eigenvectors to keep}

\item{q}{Order of penalization. Polynoms of degrees q - 1 are considered
smooth and are therefore unpenalized}

\item{verbose}{Should information about the optimization progress be
displayed}

\item{accu_dev}{Tolerance for the convergence of the optimization procedure}
}
\value{
A list containing model fit, residuals and associated uncertainty
}
\description{
Whittaker-Henderson Smoothing (Maximum Likelihood, fixed lambda)
}
